refactor: 컨트롤러 테스트 통합
